requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
-e git+https://github.com/IainMcl/BoatComp.git@60bd8bca07969c18a232d97118c6f478e9217d16#egg=boatcomp
colorama==0.4.4
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-decouple==3.4
toml==0.10.2
